requirements.txt
streamlit==1.22.0
requests==2.32.4
python-dotenv==1.2.1